Fix ScreenObject
